Desktop/eduCAS/backend_new/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
transformers
torch
sqlalchemy
psycopg2-binary
apscheduler